# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Spark Lead Hub API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /services:
    get:
      operationId: getServices
      tags: [services]
      summary: Get services
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Services list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Service"

    post:
      operationId: createService
      tags: [services]
      summary: Create service
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateServiceInput"
      responses:
        "201":
          description: Service created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Service"
  /services/{id}:
    patch:
      operationId: updateService
      tags: [services]
      summary: Update service
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateServiceInput"
      responses:
        "200":
          description: Updated service
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Service"

    delete:
      operationId: deleteService
      tags: [services]
      summary: Delete service
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /services/{id}/companies:
    get:
      operationId: getServiceCompanies
      tags: [services]
      summary: Get companies linked to a service
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Companies
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Company"

    put:
      operationId: linkServiceCompanies
      tags: [services]
      summary: Update company links for a service
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LinkCompaniesInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
  /permissions:
    get:
      operationId: getPermissions
      tags: [permissions]
      summary: Get all role permissions
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Permissions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RolePermission"

    put:
      operationId: updatePermission
      tags: [permissions]
      summary: Update a permission
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdatePermissionInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /permissions/test-email/activity:
    post:
      operationId: sendTestActivityEmail
      tags: [permissions]
      summary: Send test activity email
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /permissions/test-email/password:
    post:
      operationId: sendTestPasswordEmail
      tags: [permissions]
      summary: Send test password email
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required:
        - success
...
    Company:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        industry:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
        createdBy:
          type: string
          nullable: true
        createdAt:
          type: string
        services:
          type: array
          items:
            $ref: "#/components/schemas/ServiceBrief"
      required:
        - id
        - name
        - createdAt
...
    Service:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        category:
          type: string
          nullable: true
        description:
          type: string
          nullable: true
        createdBy:
          type: string
          nullable: true
        createdAt:
          type: string
        companies:
          type: array
          items:
            $ref: "#/components/schemas/CompanyBrief"
      required:
        - id
        - name
        - createdAt
    ServiceBrief:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
      required:
        - id
        - name
    CompanyBrief:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
      required:
        - id
        - name
    CreateServiceInput:
      type: object
      properties:
        name:
          type: string
        category:
          type: string
          nullable: true
        description:
          type: string
          nullable: true
      required:
        - name
    UpdateServiceInput:
      type: object
      properties:
        name:
          type: string
          nullable: true
        category:
          type: string
          nullable: true
        description:
          type: string
          nullable: true
    LinkCompaniesInput:
      type: object
      properties:
        companyIds:
          type: array
          items:
            type: string
      required:
        - companyIds
...
    RolePermission:
      type: object
      properties:
        id:
          type: string
        roleName:
          type: string
        resource:
          type: string
        action:
          type: string
        allowed:
          type: boolean
      required:
        - id
        - roleName
        - resource
        - action
        - allowed
    UpdatePermissionInput:
      type: object
      properties:
        roleName:
          type: string
        resource:
          type: string
        action:
          type: string
        allowed:
          type: boolean
      required:
        - roleName
        - resource
        - action
        - allowed
...