# lib/api-zod/src/generated/types/rolePermission.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Spark Lead Hub API
 * OpenAPI spec version: 0.1.0
 */

export interface RolePermission {
  id: string;
  roleName: string;
  resource: string;
  action: string;
  allowed: boolean;
}