# lib/api-zod/src/generated/types/service.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Spark Lead Hub API
 * OpenAPI spec version: 0.1.0
 */
import type { CompanyBrief } from "./companyBrief";

export interface Service {
  id: string;
  name: string;
  category?: string | null;
  description?: string | null;
  createdBy?: string | null;
  createdAt: string;
  companies?: CompanyBrief[];
}