# lib/api-zod/src/generated/types/updateServiceInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Spark Lead Hub API
 * OpenAPI spec version: 0.1.0
 */

export interface UpdateServiceInput {
  name?: string | null;
  category?: string | null;
  description?: string | null;
}